fix: removed duplicate qti-expression + qti-ordered returning correct cardinality
